+++ internal/model/comment.go
@@ -3,8 +3,8 @@ package model
 import "time"
 
 type Comment struct {
-	ID        uint      `gorm:"primaryKey" json:"id"`
-	PostID    uint      `gorm:"not null" json:"post_id"`
+	ID        uint64    `gorm:"primaryKey" json:"id"`
+	PostID    uint64    `gorm:"not null" json:"post_id"`
 	Body      string    `gorm:"type:text;not null" json:"body"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
