+++ internal/model/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID        uint      `gorm:"primaryKey" json:"id"`
+	ID        uint64    `gorm:"primaryKey" json:"id"`
 	Name      string    `gorm:"type:varchar(255);not null" json:"name"`
 	Email     string    `gorm:"unique;type:varchar(255);not null" json:"email"`
 	Password  string    `gorm:"type:varchar(255);not null" json:"password"`
@@ -13,13 +13,3 @@ type User struct {
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
-
-//func (user *User) BeforeUpdate(tx *gorm.DB) (err error) {
-//	for _, role := range user.Roles {
-//		if role.ID == 1 {
-//			return errors.New("admin user not allowed to update")
-//		}
-//	}
-//
-//	return nil
-//}
